requirements.txt
torch>=2.2
torchvision>=0.17
torchaudio>=2.2
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
scikit-learn>=1.4
opencv-python>=4.9
tqdm>=4.66
pyyaml>=6.0
